fix - fixed build erros, moved few test cases to new file to fix pmd issues
