+++ src/test/java/uk/gov/hmcts/reform/hmc/api/controllers/HearingsDataControllerTest.java
@@ -1,0 +1,154 @@
+package uk.gov.hmcts.reform.hmc.api.controllers;
+
+import org.json.simple.parser.ParseException;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.junit.runner.RunWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mockito;
+import org.mockito.MockitoAnnotations;
+import org.mockito.Spy;
+import org.mockito.junit.MockitoJUnitRunner;
+import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.test.context.ActiveProfiles;
+import uk.gov.hmcts.reform.hmc.api.model.request.HearingValues;
+import uk.gov.hmcts.reform.hmc.api.model.response.CaseHearing;
+import uk.gov.hmcts.reform.hmc.api.model.response.HearingDaySchedule;
+import uk.gov.hmcts.reform.hmc.api.model.response.ServiceHearingValues;
+import uk.gov.hmcts.reform.hmc.api.services.HearingsDataService;
+import uk.gov.hmcts.reform.hmc.api.services.IdamAuthService;
+
+import java.io.IOException;
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
+
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyString;
+import static uk.gov.hmcts.reform.hmc.api.controllers.HearingsControllerTest.feignException;
+
+@RunWith(MockitoJUnitRunner.class)
+@ExtendWith(MockitoExtension.class)
+@ActiveProfiles("test")
+public class HearingsDataControllerTest {
+
+    @InjectMocks
+    private HearingsController hearingsController;
+
+    @Spy
+    private  final IdamAuthService idamAuthService = Mockito.mock(IdamAuthService.class);
+
+    @Spy private HearingsDataService hearingsDataService;
+
+    @BeforeEach
+    void setUp() {
+
+        MockitoAnnotations.openMocks(this);
+        LocalDateTime testNextHearingDate = LocalDateTime.of(2024, 04, 28, 1, 0);
+
+        HearingDaySchedule hearingDaySchedule =
+            HearingDaySchedule.hearingDayScheduleWith()
+                .hearingVenueId("231596")
+                .hearingJudgeId("4925644")
+                .hearingStartDateTime(testNextHearingDate)
+                .build();
+        List<HearingDaySchedule> hearingDayScheduleList = new ArrayList<>();
+        hearingDayScheduleList.add(hearingDaySchedule);
+
+        CaseHearing caseHearing =
+            CaseHearing.caseHearingWith()
+                .hmcStatus("LISTED")
+                .hearingDaySchedule(hearingDayScheduleList)
+                .build();
+        List<CaseHearing> caseHearingList = new ArrayList<>();
+        caseHearingList.add(caseHearing);
+    }
+
+    @Test
+    void hearingsDataControllerTest() throws IOException, ParseException {
+
+        Mockito.when(idamAuthService.authoriseService(any())).thenReturn(Boolean.TRUE);
+
+        ServiceHearingValues hearingsData =
+            ServiceHearingValues.hearingsDataWith()
+                .hmctsServiceID("BBA3")
+                .hmctsInternalCaseName("123")
+                .publicCaseName("John Smith")
+                .caseAdditionalSecurityFlag(false)
+                .build();
+
+        Mockito.when(hearingsDataService.getCaseData(any(), anyString(), anyString()))
+            .thenReturn(hearingsData);
+
+        HearingValues hearingValues =
+            HearingValues.hearingValuesWith().hearingId("123").caseReference("123").build();
+
+        ResponseEntity<Object> hearingsData1 =
+            hearingsController.getHearingsData("Auth", "sauth", hearingValues);
+        Assertions.assertEquals(HttpStatus.OK, hearingsData1.getStatusCode());
+    }
+
+    @Test
+    void hearingsDataControllerUnauthorisedExceptionTest() throws IOException, ParseException {
+
+        HearingValues hearingValues =
+            HearingValues.hearingValuesWith().hearingId("123").caseReference("123").build();
+
+        ResponseEntity<Object> hearingsData1 =
+            hearingsController.getHearingsData("", "", hearingValues);
+
+        Assertions.assertEquals(HttpStatus.UNAUTHORIZED, hearingsData1.getStatusCode());
+    }
+
+    @Test
+    void hearingsDataControllerUnauthorisedFeignExceptionTest() throws IOException, ParseException {
+        Mockito.when(idamAuthService.authoriseService(any())).thenReturn(true);
+
+        HearingValues hearingValues =
+            HearingValues.hearingValuesWith().hearingId("123").caseReference("123").build();
+
+        Mockito.when(hearingsDataService.getCaseData(hearingValues, "", ""))
+            .thenThrow(feignException(HttpStatus.BAD_REQUEST.value(), "Not found"));
+
+        ResponseEntity<Object> hearingsData1 =
+            hearingsController.getHearingsData("", "", hearingValues);
+
+        Assertions.assertEquals(HttpStatus.BAD_REQUEST, hearingsData1.getStatusCode());
+    }
+
+    @Test
+    void hearingsDataControllerInternalServerErrorTest() throws IOException, ParseException {
+        Mockito.when(idamAuthService.authoriseService(any())).thenReturn(true);
+
+        HearingValues hearingValues =
+            HearingValues.hearingValuesWith().hearingId("123").caseReference("123").build();
+        Mockito.when(hearingsDataService.getCaseData(hearingValues, "", ""))
+            .thenThrow(new RuntimeException());
+        ResponseEntity<Object> hearingsData1 =
+            hearingsController.getHearingsData("", "", hearingValues);
+
+        Assertions.assertEquals(HttpStatus.INTERNAL_SERVER_ERROR, hearingsData1.getStatusCode());
+    }
+
+    @Test
+    void hearingsDataControllerInternalServiceErrorTest() throws IOException, ParseException {
+        Mockito.when(idamAuthService.authoriseService(any())).thenReturn(true);
+
+        HearingValues hearingValues =
+            HearingValues.hearingValuesWith().hearingId("123").caseReference("123").build();
+
+        Mockito.when(
+                hearingsDataService.getCaseData(
+                    hearingValues, "Authorization", "ServiceAuthorization"))
+            .thenThrow(feignException(HttpStatus.INTERNAL_SERVER_ERROR.value(), "Not found"));
+
+        ResponseEntity<Object> hearingsData1 =
+            hearingsController.getHearingsData("Authorization", "ServiceAuthorization", hearingValues);
+
+        Assertions.assertEquals(HttpStatus.INTERNAL_SERVER_ERROR, hearingsData1.getStatusCode());
+    }
+}
